Se eliminaron compleyamente los warnings

+++ Userland/SampleCodeModule/tests/test_processes.c
@@ -15,7 +15,7 @@ typedef struct P_rq
     enum State state;
 } p_rq;
 
-int64_t test_processes(uint64_t argc, char *argv[])
+int test_processes(int argc, char *argv[])
 {
     uint8_t rq;
     uint8_t alive = 0;
